refactor: moving the logic for modifying/creating CephCluster
(cherry picked from commit 48be5cff780bb340a9e25eb7cfbe82dc4fbedaa9)
